requirements.txt
apturl==0.5.2
asgiref==3.7.2
asn1crypto==0.24.0
asttokens==2.4.1
Brlapi==0.6.6
certifi==2018.1.18
chardet==3.0.4
click==6.7
colorama==0.3.7
command-not-found==0.3
cryptography==2.1.4
cupshelpers==1.0
decorator==5.1.1
defer==1.0.6
distlib==0.3.7
distro-info===0.18ubuntu0.18.04.1
Django==4.2
djangorestframework==3.14.0
executing==2.0.1
filelock==3.12.4
httplib2==0.9.2
idna==2.6
ipython==8.22.1
jedi==0.19.1
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
louis==3.5.0
macaroonbakery==1.1.3
Mako==1.0.7
MarkupSafe==1.0
matplotlib-inline==0.1.6
meld==3.18.0
netifaces==0.10.4
numpy==1.26.4
oauth==1.0.1
olefile==0.45.1
pandas==2.2.1
parso==0.8.3
pexpect==4.9.0
Pillow==5.1.0
pipenv==2023.10.24
platformdirs==3.11.0
prompt-toolkit==3.0.43
protobuf==3.0.0
ptyprocess==0.7.0
pure-eval==0.2.2
pycairo==1.16.2
pycodestyle==2.3.1
pycrypto==2.6.1
pycups==1.9.73
Pygments==2.17.2
PyGObject==3.26.1
pymacaroons==0.13.0
PyNaCl==1.1.2
pyRFC3339==1.0
python-apt==1.6.5+ubuntu0.7
python-dateutil==2.9.0.post0
python-debian==0.1.32
pytz==2024.1
pyxdg==0.25
PyYAML==3.12
reportlab==3.4.0
requests==2.18.4
requests-unixsocket==0.1.5
SecretStorage==2.3.1
simplejson==3.13.2
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
system-service==0.3
systemd-python==234
traitlets==5.14.1
tzdata==2024.1
ubuntu-advantage-tools==27.10
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.22
usb-creator==0.3.3
virtualenv==20.24.6
wadllib==1.3.2
wcwidth==0.2.13
xkit==0.0.0
zope.interface==4.3.2
